backend/requirements.txt
# Generated from project venv pip list (pip excluded). Pin versions for reproducibility.
# python-dotenv: load backend/.env for POSTGRES_* (not in original freeze).
python-dotenv==1.2.1
asgiref==3.11.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
curl_cffi==0.13.0
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
djangorestframework-simplejwt==5.5.1
frozendict==2.4.7
idna==3.11
multitasking==0.0.12
numpy==2.4.4
openai==1.109.1
pandas==3.0.2
peewee==4.0.4
platformdirs==4.9.4
protobuf==7.34.1
psycopg2-binary==2.9.11
pycparser==3.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
requests==2.33.1
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
websockets==16.0
yfinance==1.2.0
